add a real basic first demo
