+++ docs/src/pages/components/speed-dial/PlaygroundSpeedDial.js
@@ -1,0 +1,90 @@
+import * as React from 'react';
+import { experimentalStyled as styled } from '@material-ui/core/styles';
+import Box from '@material-ui/core/Box';
+import FormControlLabel from '@material-ui/core/FormControlLabel';
+import FormLabel from '@material-ui/core/FormLabel';
+import Radio from '@material-ui/core/Radio';
+import RadioGroup from '@material-ui/core/RadioGroup';
+import Switch from '@material-ui/core/Switch';
+import SpeedDial from '@material-ui/core/SpeedDial';
+import SpeedDialIcon from '@material-ui/core/SpeedDialIcon';
+import SpeedDialAction from '@material-ui/core/SpeedDialAction';
+import FileCopyIcon from '@material-ui/icons/FileCopyOutlined';
+import SaveIcon from '@material-ui/icons/Save';
+import PrintIcon from '@material-ui/icons/Print';
+import ShareIcon from '@material-ui/icons/Share';
+
+const StyledSpeedDial = styled(SpeedDial)(({ theme }) => ({
+  position: 'absolute',
+  '&.MuiSpeedDial-directionUp, &.MuiSpeedDial-directionLeft': {
+    bottom: theme.spacing(2),
+    right: theme.spacing(2),
+  },
+  '&.MuiSpeedDial-directionDown, &.MuiSpeedDial-directionRight': {
+    top: theme.spacing(2),
+    left: theme.spacing(2),
+  },
+}));
+
+const actions = [
+  { icon: <FileCopyIcon />, name: 'Copy' },
+  { icon: <SaveIcon />, name: 'Save' },
+  { icon: <PrintIcon />, name: 'Print' },
+  { icon: <ShareIcon />, name: 'Share' },
+];
+
+export default function PlaygroundSpeedDial() {
+  const [direction, setDirection] = React.useState('up');
+
+  const [hidden, setHidden] = React.useState(false);
+
+  const handleDirectionChange = (event) => {
+    setDirection(event.target.value);
+  };
+
+  const handleHiddenChange = (event) => {
+    setHidden(event.target.checked);
+  };
+
+  return (
+    <Box sx={{ transform: 'translateZ(0px)', flexGrow: 1 }}>
+      <FormControlLabel
+        control={
+          <Switch checked={hidden} onChange={handleHiddenChange} color="primary" />
+        }
+        label="Hidden"
+      />
+      <FormLabel sx={{ my: 1 }} component="legend">
+        Direction
+      </FormLabel>
+      <RadioGroup
+        aria-label="direction"
+        name="direction"
+        value={direction}
+        onChange={handleDirectionChange}
+        row
+      >
+        <FormControlLabel value="up" control={<Radio />} label="Up" />
+        <FormControlLabel value="right" control={<Radio />} label="Right" />
+        <FormControlLabel value="down" control={<Radio />} label="Down" />
+        <FormControlLabel value="left" control={<Radio />} label="Left" />
+      </RadioGroup>
+      <Box sx={{ position: 'relative', mt: 3, height: 320 }}>
+        <StyledSpeedDial
+          ariaLabel="SpeedDial playground example"
+          hidden={hidden}
+          icon={<SpeedDialIcon />}
+          direction={direction}
+        >
+          {actions.map((action) => (
+            <SpeedDialAction
+              key={action.name}
+              icon={action.icon}
+              tooltipTitle={action.name}
+            />
+          ))}
+        </StyledSpeedDial>
+      </Box>
+    </Box>
+  );
+}
